requirements.txt
Django==1.7.2
Fabric==1.10.1
Pillow==2.7.0
boto==2.34.0
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.19.0
django-autocomplete-light==2.0.9
django-debug-toolbar==1.2.2
django-pylibmc-sasl==0.2.4
django-recaptcha==1.0.2
django-s3-folder-storage==0.1
-e git://github.com/digi604/django-smart-selects.git@f5f7b826a543da2e6eeac37729edff9b7b8e8a64#egg=smart_selects
-e git://github.com/noirbizarre/django-eztables.git@347e74dcc08121d20f4cf942181d873dbe33b995#egg=eztables
django-storages==1.1.8
django-suit==0.2.12
django-ckeditor==4.4.7
django-tagging==0.3.4
django.js==0.8.0
djangorestframework==3.0.2
django-rest-swagger==0.2.7
django-configurations==0.8
django-cors-headers==1.0.0
django-heroku-memcacheify==0.8
ecdsa==0.11
gunicorn==19.10.0
newrelic==2.44.0.36
numpy==1.9.1
oauthlib==0.7.2
paramiko==1.15.2
psycopg2-binary==2.8.4
pycrypto==2.6.1
pylibmc==1.2.3
pystache==0.5.4
python-openid==2.2.5
redis==2.10.3
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
sorl-thumbnail==12.2
sqlparse==0.1.14
static==1.1.1
static3==0.5.1
wsgiref==0.1.2
XlsxWriter==1.2.9
